+++ CHANGELOG.rst
@@ -21,5 +21,6 @@ release (version 0.9.2)
 ----------
 
   1. Expanded flexibility for meta information access via filename for acquisition time, meas_type and filter_id in Camera class: now, the conversion / identification strings can also include the actual delimiter (e.g. delim="_", time_info_pos = 0, time_info_str="%Y%m%d_%H%M%S_%f" or filter_id_pos=3 and filter.acronym="A_dark"). This is for instance required for file naming convention of new default SO2 camera type "hd_new".
+  2. Improved functionality for dark image access in ImgList classes